pyproject.toml
[tool.poetry]
name = "small-app"
version = "0.1.0"
description = ""
authors = ["Naksen <khannanov.rus@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.103.1"}
sqlalchemy = "^2.0.20"
alembic = "^1.12.0"
asyncpg = "^0.28.0"
black = "^23.7.0"
pytest = "^7.4.2"
pytest-asyncio = "^0.21.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
ruff = "^0.0.290"
python-multipart = "^0.0.6"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [
  ".", "src",
]
asyncio_mode="auto"